Add files via upload
Loops are fun! It took me a while to really understand the concept of loops, especially for loops. Glad to say I get them now, and actually enjoy working with loops, when appropriate. This was an easy problem, but if you don't have a solid understanding of for loops, you might struggle with the syntax.